+++ archetype-pom/java-module-project/src/main/resources/archetype-resources/src/main/java/module-info.java
@@ -1,2 +1,5 @@
 module ${package} {
+    requires transitive org.apiguardian.api;
+    requires transitive org.jspecify;
+    requires transitive org.slf4j;
 }