+++ archetype-pom/java-module-project/src/test/resources/projects/it2/module/reference/src/main/java/module-info.java
@@ -1,2 +1,5 @@
 module my.pck.simple.module {
+    requires transitive org.apiguardian.api;
+    requires transitive org.jspecify;
+    requires transitive org.slf4j;
 }